Update zuul jobs: add new branches, disable old ones
Add jobs 2023.2 and 2024.1, and move to non-voting the jobs
for all unmaintained branches, but update them to reference
the correct branches (stable/foo -> unmaintained/foo).

Keep the job definitions for the older branches as they may be
referenced by zuul definitions in old branches (as this devstack
plugin is not branched).

Change-Id: I16a98ada611adf26117e66ee91645030ba2be555