zuul: run cinder-tempest-plugin tests too
Tests from cinder-tempest-plugin should work with the NFS backend
as well, so run them too.

Unfortunately it is not possible to reuse the 'full-integrated'
tox environment as it does not allow for additional regexps,
so try to mimic it by explicitly listing the tests.

This change also pins `devstack-plugin-nfs-tempest-full` to NFS
version `vers=4.2` in combination with `nconnect=4`, which improves
test stability by preventing stalls on large write operations.

[1] https://bugs.launchpad.net/ubuntu/+source/linux/+bug/2103802

Change-Id: I268f1150a5d2fd49371d84de495ff11cabc46528
Co-Authored-By: Fernando Ferraz <fernandoperches@gmail.com>
Signed-off-by: Fernando Ferraz <fernandoperches@gmail.com>